shuffle stuff around
